Initial project commit
This commit represents the initial pre-alpha version of the package
prior to any public releases of the code. This was largely worked on in
live streams, and likely requires some tweaking in order to be fully
production grade, but we need to get a version out there into peoples
hands first in order to see what else we should do.