Make editing more robust and allow key additions
This enhances the code to be more robust and tolerant of the regions
being missing or invalid in the file, and can recover and relink. We
also now properly handle joining tabs together if the setting is set
and a tab rejoins the group.

We also have the ability now to detect when CSS keys are not present and
provide the ability to add them in if they are not, so that using the
package is easier.

This requires some cleanup on the text injection which currently may use
the incorrect formatting. That will be fixed in an upcoming commit.