Clean up the text insertion code
This change keeps the same functionality as previous commit but makes
the insertion of new keys more robust (i.e. it does not assume that
the structure of the file is 100% correct, properly indented, and that
all of the keys have trailing commas).

This could be better, but this is good enough for the time being since
Sublime could gain an API in the near future to make this much more
seamless.