Resolve a plugin TypeError for Linux/MacOS
@TerminalFi reported a TypeError issue when using the plugin on MacOs in
which, while trying to create the CSS view, it bombed out instead.

After some investigation it would appear that the issue is that the
call to `sublime.load_settings()` that happens in `plugin_loaded()`
returns an invalid Settings object on Linux and MacOS but not Windows.

The result is all requests for settings returning None, even if the
setting actually exists.

This commit works around that issue by lazily loading the settings the
first time they're queried, rather than at startup time.