Fix issue with region markers getting out of sync
As originally envisioned, the code would track the total number of CSS
regions that are open in tabs in any given color scheme file so that it
could be smarter about tracking modifications that remove region
markers.

This it turns out is somewhat problematic since an undo operation can
bring a deletion region back, causing our counts to get out of sync.

Rather than trying to track the count and turn off the listener when we
think we don't need it, this just lazily creates the listener when it
is first needed (based on the first css region being created) and then
just remains in place.

This is easier and still meets the spirit of the requirement.